Move onboarding fixtures in to `mavis.test`
This moves the fixtures related to onboarding in to their own file.
