+++ mavis/test/__init__.py
@@ -1,9 +1,11 @@
 from .hooks import pytest_runtest_logreport, pytest_sessionfinish, pytest_sessionstart
 from .fixtures import (
+    admin,
     base_url,
     basic_auth,
     browser_context_args,
     children_page,
+    clinics,
     consent_page,
     dashboard_page,
     get_online_consent_url,
@@ -11,28 +13,24 @@ from .fixtures import (
     log_in_as_admin,
     log_in_as_nurse,
     log_in_page,
-    playwright_operations,
-    programmes_page,
-    school_moves_page,
-    screenshots_path,
-    sessions_page,
-    start_page,
-    test_data,
-    unmatched_page,
-    vaccines_page,
-)
-from .onboarding import (
-    admin,
-    clinics,
     nurse,
     onboard,
     onboarding,
     organisation,
+    playwright_operations,
+    programmes_page,
     reset,
+    school_moves_page,
     schools,
+    screenshots_path,
+    sessions_page,
+    start_page,
     superuser,
     team,
+    test_data,
+    unmatched_page,
     users,
+    vaccines_page,
 )
